Add user_claims_loader

+++ resources/item.py
@@ -1,6 +1,6 @@
 import sqlite3
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 
 from models.item import ItemModel
 
@@ -15,7 +15,7 @@ class Item(Resource):
 						required=True, 
 						help='Every item needs a store id')
 
-	@jwt_required()
+	@jwt_required
 	def get(self, name):
 		item = ItemModel.find_by_name(name)
 
@@ -27,7 +27,7 @@ class Item(Resource):
 		# item = next(filter(lambda x: x['name'] == name, items), None)
 		# return {'item': item}, 200 if item else 404
 
-	@jwt_required()
+	@jwt_required
 	def post(self, name):
 		if ItemModel.find_by_name(name):
 			return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -46,8 +46,12 @@ class Item(Resource):
 		
 		return item.json(), 201 # item, 201
 
-	@jwt_required()
+	@jwt_required
 	def delete(self, name):
+		claims = get_jwt_claims()
+		if not claims['is_admin']:
+			return {'message': 'Admin privilege required.'}, 401
+
 		item = ItemModel.find_by_name(name)
 		if item:
 			item.delete_from_db()
@@ -65,7 +69,7 @@ class Item(Resource):
 		# global items 
 		# items = list(filter(lambda x: x['name'] != name, items))
 
-	@jwt_required()
+	@jwt_required
 	def put(self, name):
 		data = Item.parser.parse_args()
 		item = ItemModel.find_by_name(name) # item = next(filter(lambda x: x['name'] == name, items), None)
@@ -95,7 +99,7 @@ class Item(Resource):
 
 
 class ItemList(Resource):
-	@jwt_required()
+	@jwt_required
 	def get(self):
 		return {'items': [item.json() for item in ItemModel.find_all()]}
 		# return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
